messageSizeBytes should be int

+++ sms-examples/src/main/groovy/org/aaron/sms/examples/groovy/sender/GroovySenderRunnable.groovy
@@ -14,7 +14,7 @@ class GroovySenderRunnable implements Runnable {
 
     String topicName
 
-    long messageSizeBytes
+    int messageSizeBytes
 
     long sleepBetweenSendsMS
